Add dependency on cbf-sdp-emulator-tango-device
This is an optional dependency via the "emulator" extra, so developers
of the package don't require it to be present to continue developing.
Thus this change disrupts them as little as possible.

When building the Docker image for this repository though we *do*
install the extra dependency, as it will be required when launching the
CBF with the emulator.

This work is part of YAN-702.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>