Add option to use emulated CbfSubarray
When starting the CbfSubarrayMulti server we can now indicate (via an
environment variable) whether we want to start the real CbfSubarray
device, or the emulated one. The default behavior is to start the real
device, and starting the emulated one requires having the "emulator"
extra dependency installed.

This work is part of YAN-702.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>