Add option to start emulator in Helm chart
A new midcbf.use_emulator value in the values file (defaults to false)
defines whether the emulator should be started instead of the real
CbfSubarray device. This is communicated to the device server by setting
the CBF_USE_EMULATOR environment variable to true/false, which is in
turn how the device server starts one device or the other.

This work is part of YAN-702.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>